requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
Markdown==3.7
psycopg2-binary==2.9.10
PyJWT==2.10.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.8.2
